backend/pyproject.toml
[project]
name = "civo-backend"
version = "0.1.0"
description = "Civo — MA permitting intelligence backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "pydantic>=2.5",
    "sqlalchemy>=2.0",
    "geoalchemy2>=0.15",
    "alembic>=1.13",
    "psycopg2-binary>=2.9",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "anthropic>=0.39",
    "shapely>=2.0",
    "python-dotenv>=1.0",
    "reportlab>=4.0",
    "lightgbm>=4.3",
    "pandas>=2.0",
    "scikit-learn>=1.4",
    "numpy>=1.26",
    "Pillow>=10.0",
    "pdfplumber>=0.10",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "openpyxl>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.10",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "ingest*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
